+++ incubator/polyfills/package.json
@@ -10,7 +10,6 @@
     "email": "microsoftopensource@users.noreply.github.com"
   },
   "files": [
-    "babel-plugin.js",
     "lib/*"
   ],
   "main": "lib/index.js",
